requirements.txt
Django==2.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-compressor==2.2
pytz==2017.3
rcssmin==1.0.6
rjsmin==1.0.12
